+++ yuyo/list_status.py
@@ -788,7 +788,7 @@ class ServiceManager(AbstractManager):
 
     async def _loop(self) -> None:
         # This acts as a priority queue.
-        queue = [(service.repeat, service) for service in self._services]
+        queue: list[tuple[float, _ServiceDescriptor]] = [(service.repeat, service) for service in self._services]
         while True:
             await asyncio.sleep(sleep := queue[0][0])
             queue = [(time_ - sleep, service) for time_, service in queue]
@@ -1001,7 +1001,6 @@ class DiscordBotListService:
         ) as response:
             if shard_id is None:
                 await _log_response("Discordbotlist.com", response, is_global=False)
-                return
 
             if response.status in _RETRY_ERROR_CODES:
                 if retry_after := response.headers.get(_RETRY_AFTER_KEY):
@@ -1010,3 +1009,5 @@ class DiscordBotListService:
                 return -1
 
             response.raise_for_status()
+
+        return None  # MyPy compatibility